requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
sentence-transformers
chromadb==0.5.3
requests
pydantic
streamlit
scikit-learn
joblib
acryl-datahub
datahub
# Caso esiver usando o banco postgress acryl-datahub[postgres]